requirements.txt
# Azure AI Integration Dependencies

# Azure OpenAI SDK
openai>=1.0.0

# Environment variable management
python-dotenv>=1.0.0

# HTTP client for REST API calls (Translator)
requests>=2.31.0

# Flask Web Framework
flask>=3.0.0

# Document Processing
pdfplumber>=0.10.0
python-docx>=1.0.0

# File upload handling
werkzeug>=3.0.0
